Revert "match test types to prevent Windows from failing tests"
This reverts commit a990307066e99cb56e15ae5de63e67e29cc72911.

+++ dlib/test/rand.cpp
@@ -475,12 +475,12 @@ namespace
 
         {
             dlib::rand rnd;
-            std::vector<uint32> out;
+            std::vector<long> out;
             for (int i = 0; i < 30; ++i) {
                 out.push_back(rnd.get_random_32bit_number());
             }
 
-            const std::vector<uint32> expected = {
+            const std::vector<long> expected = {
                 725333953,251387296,3200466189,2466988778,2049276419,2620437198,2806522923,
                 2922190659,4151412029,2894696296,1344442829,1165961100,328304965,1533685458,
                 3399102146,3995382051,1569312238,2353373514,2512982725,2903494783,787425157,
@@ -492,12 +492,12 @@ namespace
         {
             dlib::rand rnd;
             rnd.set_seed("this seed");
-            std::vector<uint32> out;
+            std::vector<long> out;
             for (int i = 0; i < 30; ++i) {
                 out.push_back(rnd.get_random_32bit_number());
             }
 
-            const std::vector<uint32> expected = {
+            const std::vector<long> expected = {
                 856663397,2356564049,1192662566,3478257893,1069117227,
                 1922448468,497418632,2504525324,987414451,769612124,77224022,2998161761,
                 1364481427,639342008,1778351952,1931573847,3213816676,3019312695,4179936779,
@@ -509,12 +509,12 @@ namespace
         {
             dlib::rand rnd;
             rnd.set_seed("some other seed");
-            std::vector<long long> out;
+            std::vector<long> out;
             for (int i = 0; i < 30; ++i) {
                 out.push_back(rnd.get_integer(1000));
             }
 
-            const std::vector<long long> expected = {
+            const std::vector<long> expected = {
                 243,556,158,256,772,84,837,920,767,769,939,394,121,367,575,877,861,506,
                 451,845,870,638,825,516,327,25,646,373,386,227};
             DLIB_TEST(out == expected);
